finally rename namespace with proper name
