+++ tests/Nexucis/Tests/Elasticsearch/Helper/Nodowntime/Parameter/SearchParameterTest.php
@@ -2,7 +2,7 @@
 
 namespace Nexucis\Tests\Elasticsearch\Helper\Nodowntime\Parameter;
 
-use Nexucis\Elasticsearch\Helper\Nodowntime\DumpNamespace\SearchParameter;
+use Nexucis\Elasticsearch\Helper\Nodowntime\Parameter\SearchParameter;
 use PHPUnit\Framework\TestCase;
 
 class SearchParameterTest extends TestCase
